+++ admin/template/user.php
@@ -83,48 +83,56 @@ if ($page1 == "e" || $page1 == "edp" || $page1 == "ene") { ?>
             <th><?php echo $tl["menu"]["m9"]; ?></th>
             <th></th>
             <th>
-              <button type="submit" name="password" id="button_key" class="btn btn-default btn-xs"
-                      onclick="if(!confirm('<?php echo $tl["user"]["pass"]; ?>'))return false;"><i
-                  class="fa fa-key"></i></button>
+              <button type="submit" name="password" id="button_key" class="btn btn-default btn-xs" onclick="if(!confirm('<?php echo $tl["user"]["pass"]; ?>'))return false;">
+                <i class="fa fa-key"></i>
+              </button>
             </th>
             <th></th>
             <th>
-              <button type="submit" name="delete" id="button_delete" class="btn btn-danger btn-xs"
-                      onclick="if(!confirm('<?php echo $tl["user"]["al"]; ?>'))return false;"><i
-                  class="fa fa-trash-o"></i></button>
+              <button type="submit" name="delete" id="button_delete" class="btn btn-danger btn-xs" onclick="if(!confirm('<?php echo $tl["user"]["al"]; ?>'))return false;">
+                <i class="fa fa-trash-o"></i>
+              </button>
             </th>
           </tr>
           </thead>
           <?php foreach ($JAK_USER_ALL_APPROVE as $va) { ?>
             <tr>
               <td><?php echo $va["id"]; ?></td>
-              <td><input type="checkbox" name="jak_delete_user[]" class="highlight" value="<?php echo $va["id"]; ?>"/>
+              <td>
+                <input type="checkbox" name="jak_delete_user[]" class="highlight" value="<?php echo $va["id"]; ?>"/>
               </td>
-              <td><a
-                  href="index.php?p=user&amp;sp=edit&amp;ssp=<?php echo $va["id"]; ?>"><?php echo $va["username"]; ?></a>
+              <td>
+                <a href="index.php?p=user&amp;sp=edit&amp;ssp=<?php echo $va["id"]; ?>"><?php echo $va["username"]; ?></a>
               </td>
-              <td><a
-                  href="index.php?p=user&amp;sp=edit&amp;ssp=<?php echo $va["id"]; ?>"><?php echo $va["email"]; ?></a>
+              <td>
+                <a href="index.php?p=user&amp;sp=edit&amp;ssp=<?php echo $va["id"]; ?>"><?php echo $va["email"]; ?></a>
               </td>
-              <td><?php if (isset($JAK_USERGROUP_ALL) && is_array($JAK_USERGROUP_ALL)) foreach ($JAK_USERGROUP_ALL as $z) {
+              <td>
+                <?php if (isset($JAK_USERGROUP_ALL) && is_array($JAK_USERGROUP_ALL)) foreach ($JAK_USERGROUP_ALL as $z) {
                   if ($va["usergroupid"] == $z["id"]) { ?><a
                     href="index.php?p=usergroup&amp;sp=user&amp;ssp=<?php echo $z["id"]; ?>"><?php echo $z["name"]; ?></a><?php }
-                } ?></td>
-              <td class="content-go"><a href="index.php?p=user&amp;sp=verify&amp;ssp=<?php echo $va["id"]; ?>"
-                                        class="btn btn-default btn-xs"><i
-                    class="fa fa-<?php if ($va["access"] == 3 || $va["access"] == 2) { ?>envelope-o<?php } else { ?>lock<?php } ?>"></i></a>
+                } ?>
               </td>
-              <td><a class="btn btn-default btn-xs"
-                     href="index.php?p=user&amp;sp=password&amp;ssp=<?php echo $va["id"]; ?>"
-                     onclick="if(!confirm('<?php echo $tl["user"]["pass"]; ?>'))return false;"><i class="fa fa-key"></i></a>
+              <td class="content-go">
+                <a href="index.php?p=user&amp;sp=verify&amp;ssp=<?php echo $va["id"]; ?>" class="btn btn-default btn-xs">
+                  <i class="fa fa-<?php if ($va["access"] == 3 || $va["access"] == 2) { ?>envelope-o<?php } else { ?>lock<?php } ?>"></i>
+                </a>
               </td>
-              <td><a class="btn btn-default btn-xs"
-                     href="index.php?p=user&amp;sp=edit&amp;ssp=<?php echo $va["id"]; ?>"><i class="fa fa-edit"></i></a>
+              <td>
+                <a class="btn btn-default btn-xs" href="index.php?p=user&amp;sp=password&amp;ssp=<?php echo $va["id"]; ?>" onclick="if(!confirm('<?php echo $tl["user"]["pass"]; ?>'))return false;">
+                  <i class="fa fa-key"></i>
+                </a>
               </td>
-              <td><a class="btn btn-default btn-xs"
-                     href="index.php?p=user&amp;sp=delete&amp;ssp=<?php echo $va["id"]; ?>"
-                     onclick="if(!confirm('<?php echo $tl["user"]["al"]; ?>'))return false;"><i
-                    class="fa fa-trash-o"></i></a></td>
+              <td>
+                <a class="btn btn-default btn-xs" href="index.php?p=user&amp;sp=edit&amp;ssp=<?php echo $va["id"]; ?>">
+                  <i class="fa fa-edit"></i>
+                </a>
+              </td>
+              <td>
+                <a class="btn btn-default btn-xs" href="index.php?p=user&amp;sp=delete&amp;ssp=<?php echo $va["id"]; ?>" onclick="if(!confirm('<?php echo $tl["user"]["al"]; ?>'))return false;">
+                  <i class="fa fa-trash-o"></i>
+                </a>
+              </td>
             </tr>
           <?php } ?>
         </table>
@@ -141,60 +149,81 @@ if ($page1 == "e" || $page1 == "edp" || $page1 == "ene") { ?>
           <tr>
             <th>#</th>
             <th><input type="checkbox" id="jak_delete_all"/></th>
-            <th><?php echo $tl["user"]["u2"]; ?> <a class="btn btn-warning btn-xs"
-                                                    href="index.php?p=user&amp;sp=sort&amp;ssp=username&amp;sssp=DESC"><i
-                  class="fa fa-arrow-up"></i></a> <a class="btn btn-success btn-xs"
-                                                     href="index.php?p=user&amp;sp=sort&amp;ssp=username&amp;sssp=ASC"><i
-                  class="fa fa-arrow-down"></i></a></th>
-            <th><?php echo $tl["user"]["u1"]; ?> <a class="btn btn-warning btn-xs"
-                                                    href="index.php?p=user&amp;sp=sort&amp;ssp=email&amp;sssp=DESC"><i
-                  class="fa fa-arrow-up"></i></a> <a class="btn btn-success btn-xs"
-                                                     href="index.php?p=user&amp;sp=sort&amp;ssp=email&amp;sssp=ASC"><i
-                  class="fa fa-arrow-down"></i></a></th>
-            <th><?php echo $tl["menu"]["m9"]; ?></th>
             <th>
-              <button type="submit" name="lock" id="button_lock" class="btn btn-default btn-xs"><i
-                  class="fa fa-lock"></i></button>
+              <?php echo $tl["user"]["u2"]; ?>
+              <a class="btn btn-warning btn-xs" href="index.php?p=user&amp;sp=sort&amp;ssp=username&amp;sssp=DESC">
+                <i class="fa fa-arrow-up"></i>
+              </a>
+              <a class="btn btn-success btn-xs" href="index.php?p=user&amp;sp=sort&amp;ssp=username&amp;sssp=ASC">
+                <i class="fa fa-arrow-down"></i>
+              </a>
             </th>
             <th>
-              <button type="submit" name="password" id="button_key" class="btn btn-default btn-xs"
-                      onclick="if(!confirm('<?php echo $tl["user"]["pass"]; ?>'))return false;"><i
-                  class="fa fa-key"></i></button>
+              <?php echo $tl["user"]["u1"]; ?>
+              <a class="btn btn-warning btn-xs" href="index.php?p=user&amp;sp=sort&amp;ssp=email&amp;sssp=DESC">
+                <i class="fa fa-arrow-up"></i>
+              </a>
+              <a class="btn btn-success btn-xs" href="index.php?p=user&amp;sp=sort&amp;ssp=email&amp;sssp=ASC">
+                <i class="fa fa-arrow-down"></i>
+              </a>
+            </th>
+            <th><?php echo $tl["menu"]["m9"]; ?></th>
+            <th>
+              <button type="submit" name="lock" id="button_lock" class="btn btn-default btn-xs">
+                <i class="fa fa-lock"></i>
+              </button>
+            </th>
+            <th>
+              <button type="submit" name="password" id="button_key" class="btn btn-default btn-xs" onclick="if(!confirm('<?php echo $tl["user"]["pass"]; ?>'))return false;">
+                <i class="fa fa-key"></i>
+              </button>
             </th>
             <th></th>
             <th>
-              <button type="submit" name="delete" id="button_delete" class="btn btn-danger btn-xs"
-                      onclick="if(!confirm('<?php echo $tl["user"]["al"]; ?>'))return false;"><i
-                  class="fa fa-trash-o"></i></button>
+              <button type="submit" name="delete" id="button_delete" class="btn btn-danger btn-xs" onclick="if(!confirm('<?php echo $tl["user"]["al"]; ?>'))return false;">
+                <i class="fa fa-trash-o"></i>
+              </button>
             </th>
           </tr>
           </thead>
           <?php if (isset($JAK_USER_ALL) && is_array($JAK_USER_ALL)) foreach ($JAK_USER_ALL as $v) { ?>
             <tr>
               <td><?php echo $v["id"]; ?></td>
-              <td><input type="checkbox" name="jak_delete_user[]" class="highlight" value="<?php echo $v["id"]; ?>"/>
+              <td>
+                <input type="checkbox" name="jak_delete_user[]" class="highlight" value="<?php echo $v["id"]; ?>"/>
               </td>
-              <td><a
-                  href="index.php?p=user&amp;sp=edit&amp;ssp=<?php echo $v["id"]; ?>"><?php echo $v["username"]; ?></a>
+              <td>
+                <a href="index.php?p=user&amp;sp=edit&amp;ssp=<?php echo $v["id"]; ?>"><?php echo $v["username"]; ?></a>
               </td>
-              <td><a href="index.php?p=user&amp;sp=edit&amp;ssp=<?php echo $v["id"]; ?>"><?php echo $v["email"]; ?></a>
+              <td>
+                <a href="index.php?p=user&amp;sp=edit&amp;ssp=<?php echo $v["id"]; ?>"><?php echo $v["email"]; ?></a>
               </td>
-              <td><?php if (isset($JAK_USERGROUP_ALL) && is_array($JAK_USERGROUP_ALL)) foreach ($JAK_USERGROUP_ALL as $z) {
+              <td>
+                <?php if (isset($JAK_USERGROUP_ALL) && is_array($JAK_USERGROUP_ALL)) foreach ($JAK_USERGROUP_ALL as $z) {
                   if ($v["usergroupid"] == $z["id"]) { ?><a
                     href="index.php?p=usergroup&amp;sp=user&amp;ssp=<?php echo $z["id"]; ?>"><?php echo $z["name"]; ?></a><?php }
-                } ?></td>
-              <td><a class="btn btn-default btn-xs" href="index.php?p=user&amp;sp=lock&amp;ssp=<?php echo $v["id"]; ?>"><i
-                    class="fa fa-<?php if ($v["access"] == '1') { ?>check<?php } else { ?>lock<?php } ?>"></i></a></td>
-              <td><a class="btn btn-default btn-xs"
-                     href="index.php?p=user&amp;sp=password&amp;ssp=<?php echo $v["id"]; ?>"
-                     onclick="if(!confirm('<?php echo $tl["user"]["pass"]; ?>'))return false;"><i class="fa fa-key"></i></a>
+                } ?>
               </td>
-              <td><a class="btn btn-default btn-xs" href="index.php?p=user&amp;sp=edit&amp;ssp=<?php echo $v["id"]; ?>"><i
-                    class="fa fa-edit"></i></a></td>
-              <td><a class="btn btn-default btn-xs"
-                     href="index.php?p=user&amp;sp=delete&amp;ssp=<?php echo $v["id"]; ?>"
-                     onclick="if(!confirm('<?php echo $tl["user"]["al"]; ?>'))return false;"><i
-                    class="fa fa-trash-o"></i></a></td>
+              <td>
+                <a class="btn btn-default btn-xs" href="index.php?p=user&amp;sp=lock&amp;ssp=<?php echo $v["id"]; ?>" data-toggle="tooltip" data-placement="bottom" title="<?php if ($v["access"] == '1') { echo $tl["icons"]["i6"]; } else { echo $tl["icons"]["i5"]; } ?>">
+                  <i class="fa fa-<?php if ($v["access"] == '1') { ?>check<?php } else { ?>lock<?php } ?>"></i>
+                </a>
+              </td>
+              <td>
+                <a class="btn btn-default btn-xs" href="index.php?p=user&amp;sp=password&amp;ssp=<?php echo $v["id"]; ?>" onclick="if(!confirm('<?php echo $tl["user"]["pass"]; ?>'))return false;" data-toggle="tooltip" data-placement="bottom" title="<?php echo $tl["icons"]["i14"]; ?>">
+                  <i class="fa fa-key"></i>
+                </a>
+              </td>
+              <td>
+                <a class="btn btn-default btn-xs" href="index.php?p=user&amp;sp=edit&amp;ssp=<?php echo $v["id"]; ?>" data-toggle="tooltip" data-placement="bottom" title="<?php echo $tl["icons"]["i2"]; ?>">
+                  <i class="fa fa-edit"></i>
+                </a>
+              </td>
+              <td>
+                <a class="btn btn-default btn-xs" href="index.php?p=user&amp;sp=delete&amp;ssp=<?php echo $v["id"]; ?>" onclick="if(!confirm('<?php echo $tl["user"]["al"]; ?>'))return false;" data-toggle="tooltip" data-placement="bottom" title="<?php echo $tl["icons"]["i1"]; ?>">
+                  <i class="fa fa-trash-o"></i>
+                </a>
+              </td>
             </tr>
           <?php } ?>
         </table>
